Backend/requirements.txt
annotated-types
anyio
assemblyai
asttokens
bcrypt==4.0.1
beautifulsoup4
cachetools
certifi
cffi
charset-normalizer
click>=8.0.0
colorama
comtypes
cryptography
decorator
distro
dnspython
docopt
ecdsa
elevenlabs==0.3.0b0
email_validator
executing
fastapi
filelock
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-speech
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
jedi
Jinja2
jiter
Js2Py
keyboard
llvmlite
MarkupSafe
matplotlib-inline
more-itertools
motor
mpmath
numba
numpy
openai
packaging
parso
passlib
pipwin
prompt_toolkit
proto-plus
protobuf
pure_eval
pyasn1
pyasn1_modules
pycparser
pydantic>=2.0.0
pydantic_core
Pygments
pyjsparser
pymongo
pyparsing
PyPrind
pySmartDL
python-dotenv
python-jose
python-multipart
pyttsx3
regex
requests
rsa
scipy
setuptools
six
sniffio
sounddevice
soupsieve
SpeechRecognition
stack-data
starlette
sympy
tiktoken
tqdm
traitlets
typing-inspection
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
uvicorn
wcwidth
websockets
razorpay
yagmail





